+++ view/frontend/templates/paypal/product_page.phtml
@@ -4,7 +4,7 @@
  */
 $id = $block->getContainerId() . mt_rand();
 $config = [
-    'Magento_Braintree/js/paypal/one-click-button' => [
+    'Magento_Braintree/js/paypal/product-page' => [
         'id' => $id,
         'environment' => $block->getEnvironment(),
         'clientToken' => $block->getClientToken(),
