+++ view/frontend/web/js/paypal/product-page.js
