+++ view/frontend/web/js/view/product-page.js
